src/requirements.txt
opencv-python
scipy
python-socketio
eventlet
requests
flask-socketio
pyzbar